design: introduce를 about으로 경로, 파일명 변경
