+++ frontend/src/components/about/FunctionSection.tsx
