+++ frontend/src/components/about/MemberCard.tsx
