+++ frontend/src/components/about/TeamSection.tsx
