+++ frontend/src/components/about/WelcomeSection.tsx
